Extract colors into shared object
