+++ app/javascript/shared/colors.js
@@ -1,0 +1,10 @@
+const colors = {
+  background: "#800080",
+  backgroundHover: "#670067",
+  backgroundDisabled: "#4d004d",
+  error: "#ffff00",
+  pageBackground: "#111",
+  text: "#eee"
+}
+
+export default colors
